+++ cli/utils/crossChain.ts
@@ -41,25 +41,25 @@ export async function checkCrossChainSupport(
     );
 
     if (
-      promptYesNo(
+      await promptYesNo(
         `${colors.yellow}Do you want to add Hyperlane data? (y/n):${colors.reset}`
       )
     ) {
-      auxChainData.Hyperlane.DomainId = promptNumber(
+      auxChainData.Hyperlane.DomainId = await promptNumber(
         `${colors.yellow}Enter Hyperlane Domain ID for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
-      auxChainData.Hyperlane.MailboxAddress = promptAddress(
+      auxChainData.Hyperlane.MailboxAddress = await promptAddress(
         `${colors.yellow}Enter Hyperlane Mailbox Address for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
     } else {
       if (
-        !promptYesNo(
+        !(await promptYesNo(
           `${colors.yellow}Do you want to continue without adding Hyperlane data? (y/n):${colors.reset}`
-        )
+        ))
       ) {
         criticalErrorCustom(
           `User opted to not add Hyperlane data.`,
@@ -80,25 +80,25 @@ export async function checkCrossChainSupport(
     );
 
     if (
-      promptYesNo(
+      await promptYesNo(
         `${colors.yellow}Do you want to add LayerZero data? (y/n):${colors.reset}`
       )
     ) {
-      auxChainData.LayerZero.EId = promptNumber(
+      auxChainData.LayerZero.EId = await promptNumber(
         `${colors.yellow}Enter LayerZero EId for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
-      auxChainData.LayerZero.EndpointAddress = promptAddress(
+      auxChainData.LayerZero.EndpointAddress = await promptAddress(
         `${colors.yellow}Enter LayerZero Endpoint Address for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
     } else {
       if (
-        !promptYesNo(
+        !(await promptYesNo(
           `${colors.yellow}Do you want to continue without adding LayerZero data? (y/n):${colors.reset}`
-        )
+        ))
       ) {
         criticalErrorCustom(
           `User opted to not add LayerZero data.`,
@@ -119,30 +119,30 @@ export async function checkCrossChainSupport(
     );
 
     if (
-      promptYesNo(
+      await promptYesNo(
         `${colors.yellow}Do you want to add Axelar data? (y/n):${colors.reset}`
       )
     ) {
-      auxChainData.Axelar.ChainName = promptString(
+      auxChainData.Axelar.ChainName = await promptString(
         `${colors.yellow}Enter Axelar Chain Name for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
-      auxChainData.Axelar.Gateway = promptAddress(
+      auxChainData.Axelar.Gateway = await promptAddress(
         `${colors.yellow}Enter Axelar Gateway Address for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
-      auxChainData.Axelar.GasService = promptAddress(
+      auxChainData.Axelar.GasService = await promptAddress(
         `${colors.yellow}Enter Axelar Gas Service Address for ${
           chainData!.Chain
         } (${chainId}):${colors.reset} `
       );
     } else {
       if (
-        !promptYesNo(
+        !(await promptYesNo(
           `${colors.yellow}Do you want to continue without adding Axelar data? (y/n):${colors.reset}`
-        )
+        ))
       ) {
         criticalErrorCustom(
           `User opted to not add Axelar data.`,